requirement.txt
keras==2.13.1
numpy==1.24.3
tensorflow==2.13.0
pillow==9.5.0
boto3==1.28.30
PyMuPDF==1.22.5
flask==2.3.2